Remove debug logs

+++ scoreboard.cpp
@@ -24,7 +24,7 @@ void Scoreboard::reset_order(void) {
 }
 
 vector<int> Scoreboard::get_by_rank(int start_index, int end_index) {
-	end_index = min(end_index, this->teams.size());
+	end_index = min(end_index, (int)this->teams.size());
 	if (end_index <= start_index or start_index >= this->teams.size())
 		return vector<int>();
 	return vector<int>(this->teams.begin() + start_index, this->teams.begin() + end_index);
